+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 - Unreleased
+
+### Added
+- Assign reviewers using labels
+
+### Changed
+- [breaking change] `ignore_files` renamed to `ignore`
+
 ## 1.0.1 - 26.08.2021
 
 ### Changed